changed the script function interface
Script functions now return a list of scene computations (effectively one
atomic state of the script). The main function will now get the list of
script computations from the Script class function `run`, and send that
to be executed by the loop function, which will react to input between
each state of the script.