removed all string manipulation for component JSON creation
Got rid of all the playing with strings and concatenation functions to
use just plain old JSON serialization and component types. But you could
still specify all the data in a JSON file and use that string for
component creation. The JSON types used in the program are just for ease
of creation while a file loading mechanism is not in place.