removed networking, separate event modules
Before the events were separated into different modules and ADTs which
makes sense in an OO lang to keep things decoupled, but this makes it
hard to have a collection of generic events. The events are now all in
one ADT, but this is probably temporary.

Also removed the InstanceState export dependency in Main.hs. This allows me
to restructure the type without having to go all over the place to fix
changes. All the changes will be kept in the Instance module.