removed character damage attribute
The character component damage attribute was dropped, and now there is
an equipment attribute which only includes a weapon for now.

Also added a unit test to make sure characters are getting deleted after
their health goes below zero. Every update the characterManager will
check its components to ensure there are none with a health below zero.
If there are the manager will push a DeathEvent and then delete the
component during the next update.