added a new ai function: follow
Follow will copy any movement the player makes.

Also fixed the ai manager update so that each GOiD does NOT do a
computation for every type of ai computer. Before I was using the
applicative <*> to concatentate a list of GOiDs and a list of ai
computers. The result was a list of each goid and the computation for
each computer (not what I wanted). Now it has been changed so that the
update function actually does what one would expect.

Some minor refactoring.